Shows probability and rating
