--- heroku/temp.py
@@ -1,6 +1,0 @@
-from tonami import pitch_process as pp
-
-def main():
-    # pd.save_classifier_data()
-    print("hello")
-main()
